refactor TestsBatch
