+++ lib/knapsack_pro.rb
@@ -81,6 +81,7 @@ require_relative 'knapsack_pro/runners/queue/cucumber_runner'
 require_relative 'knapsack_pro/runners/queue/minitest_runner'
 require_relative 'knapsack_pro/store/server'
 require_relative 'knapsack_pro/store/queue_batch_manager'
+require_relative 'knapsack_pro/store/tests_batch'
 require_relative 'knapsack_pro/test_case_detectors/rspec_test_example_detector'
 require_relative 'knapsack_pro/crypto/encryptor'
 require_relative 'knapsack_pro/crypto/branch_encryptor'
